requirements.txt
psycopg2-binary
flasgger
sqlalchemy[asyncio]
asyncpg
flask-swag
swagger-ui-py
pytest
pytest-asyncio
httpx
sqlalchemy
quart
Flask
Flask-SQLAlchemy
hypercorn
